modify math for norm
